+++ tests/games/games.integration.test.ts
@@ -2,7 +2,7 @@ import request from 'supertest';
 import express from 'express';
 import gameRoutes from '../../src/features/games/gameRoutes';
 import connectDB from '../../src/config/db';
-import { mockBuiltGames } from '../__mocks__/mock-built-games';
+import { mockBuiltGames } from '../__mocks__/mockBuiltGames';
 import Game from '../../src/features/games/gameModel';
 import dotenv from 'dotenv';
 dotenv.config({ path: '.env.test' });
